move around flatpak utility functions and implement workaround for flatpak bug
until https://github.com/flatpak/flatpak/pull/6084 is merged and available
in distributions
